[Feature-MainHomeApi] Add Schedule Ui Model
